The middleware
Middleware can install profiles in BigSur with Munki, VPP apps with pkginfo files, enable ARD, etc. 

The purpose of the middleware is to control what gets exposed from the munki repo, you can update the user:password key value pair in the env on a schedule and you can limit what commands go through, for example I can't send an EraseDevice with it because it is not a supported command.